Initial implementation using kadi to serve telemetry
